% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entomological_data}
\alias{entomological_data}
\title{Example dataset containing abundances of mosquitoes vectors of malaria. Used in article 'use_case'.}
\format{
## `entomological_data`
A data frame with 232 rows and 6 columns:
\describe{
  \item{mission}{number of the entomological survey}
  \item{date}{date of the survey}
  \item{village}{3-digit code for the village of the survey}
  \item{X, Y}{longitude and latitude of the center of the village}
  \item{n}{number of mosquitoes collected}
}
}
\source{
<https://doi.org/10.15468/v8fvyn>
}
\usage{
entomological_data
}
\description{
Example dataset containing abundances of mosquitoes vectors of malaria. Used in article 'use_case'.
}
\keyword{datasets}
