% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection_info.R
\name{get_collection_info}
\alias{get_collection_info}
\title{Get information on a collection of interest}
\usage{
get_collection_info(collection)
}
\arguments{
\item{collection}{string. Collection of interest.}
}
\value{
Opens a browser window with the metadata of the collection (points to the DOI of the collection)
}
\description{
Opens a browser window with the metadata of the collection
}
\examples{

# Get the collections implemented :

(get_collection_info("MOD11A1.006"))

}
