% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optional_parameters.R
\name{get_optional_parameters}
\alias{get_optional_parameters}
\title{Precompute the parameter \code{opt_param} of the function \link{get_url}}
\usage{
get_optional_parameters(collection, roi, login_credentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{roi}{object of class \code{sf} or \code{sfc} Region of interest. Must be POLYGON-type geometry. Can be composed of several features (see details)}

\item{login_credentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
a list with the following named objects :
\itemize{
 \item{*roiSpatialIndexBound*: }{OPeNDAP indices for the spatial coordinates of the bounding box of the ROI (minLat, maxLat, minLon, maxLon)}
 \item{*availableVariables*: }{Variables available for the collection of interest}
 \item{*roiSpatialBound*: }{The spatial coordinates of the bounding box of the ROI expressed in the CRS of the collection}
 \item{*OpenDAPXVector*: }{The X (longitude) vector}
 \item{*OpenDAPYVector*: }{The Y (longitude) vector}
 \item{*OpenDAPtimeVector*: }{The time vector, or NULL if the collection does not have a time vector}
 \item{*modis_tile*: }{The MODIS tile(s) number(s) for the ROI or NULL if the collection is not MODIS}
}
}
\description{
Precompute the parameter \code{opt_param} to further provide as input of the \link{get_url} function. Useful to speed-up the overall processing time.
}
\details{
When it is needed to loop the function \link{get_url} over several time frames, it is advised to previously run the function \code{get_optional_parameters} and provide the output as input \code{opt_param} parameter of the \link{get_url} function.
This will save much time, as internal parameters will be calculated only once.
}
\examples{

\dontrun{
require(sf)

# Login to Earthdata
earthdata_credentials<-readLines("/home/ptaconet/opendapr/.earthdata_credentials.txt")
username=strsplit(earthdata_credentials,"=")[[1]][2]
password=strsplit(earthdata_credentials,"=")[[2]][2]
login<-login_usgs(c(username,password))

# Get the optional parameters for the collection MOD11A1.006 and the roi :
roi <- sf::st_read(system.file("extdata/roi_example.gpkg", package = "opendapr"),quiet=TRUE)
(opt_param_mod11a1 <- get_optional_parameters("MOD11A1.006",roi) )

}
}
