% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_info.R
\name{get_variables_info}
\alias{get_variables_info}
\title{Get variables/dimensions information for a collection of interest}
\usage{
get_variables_info(collection, loginCredentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
A data.frame with the available variables for the collection, and a set of related information for each variable.
These variables can be provided as input parameter \code{variables} of the function \link{get_url}
}
\description{
Outputs a data frame with the variables available for a given collection
}
\examples{

\dontrun{
# login to earthdata
earthdata_credentials<-readLines("/home/ptaconet/opendapr/.earthdata_credentials.txt")
earthdata_username=strsplit(earthdata_credentials,"=")[[1]][2]
earthdata_password=strsplit(earthdata_credentials,"=")[[2]][2]
login<-login_earthdata(c(earthdata_username,earthdata_password))

# Get the collections implemented in the package :
collections_available <- get_collections_available()

# Get the variables available for the collection MOD11A1.006
(df_varinfo<-get_variables_info("MOD11A1.006"))
}

}
