% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{getMODIStileNames}
\alias{getMODIStileNames}
\title{Get MODIS tile(s) intersecting a given ROI}
\usage{
getMODIStileNames(roi)
}
\arguments{
\item{roi}{object of class \code{sf} or \code{sfc}. Region of interest}
}
\value{
a character string vector with the MODIS tiles intersecting the ROI
}
\description{
Get MODIS tile(s) intersecting a given ROI
}
\details{
If the ROI is covering multiple MODIS tiles, consider splitting the ROI into multiple parts (1 for each MODIS tile) before executing the \code{getUrl} function.
Then loop \code{getUrl} over each part. Check out the example for some tips on how to proceed.
}
\note{
The function uses the MODIS tile gridding system available at https://modis.ornl.gov/files/modis_sin.kmz . The same dataset is available in the package through opendapr::modis_tiles
}
\examples{

require(sf)
require(purrr)

## ROI intersecting one MODIS tile
roi <- st_read(system.file("extdata/roi_example.gpkg", package = "opendapr"),quiet=TRUE)
(modis_tiles <- getMODIStileNames(roi))

## ROI intersecting 2 MODIS tiles (works for n MODIS tiles)
roi <- st_read(system.file("extdata/modis2tiles.gpkg", package = "opendapr"),quiet=TRUE)
(modis_tiles <- getMODIStileNames(roi))

# Split the ROI into the various MODIS tiles
roi_div <- roi \%>\%
st_intersection(opendapr:::modis_tiles) \%>\%
split(f = seq(nrow(.)))

# Login to earthdata
path_to_earthdata_credentials <- "/home/ptaconet/react/datasets/credentials_earthdata.txt"
earthdata_credentials<-readLines(path_to_earthdata_credentials)
earthdata_username<-strsplit(earthdata_credentials,"=")[[1]][2]
earthdata_password<-strsplit(earthdata_credentials,"=")[[2]][2]
login_earthdata(c(earthdata_username,earthdata_password))

# Get the URLs
odap_urls_mod11a1 <- purrr::map_dfr(roi_div,~getUrl(
collection = "MOD11A1.006",
variables = c("LST_Day_1km","LST_Night_1km"),
roi = .,
timeRange = as.Date(c("2017-01-01","2017-01-30"))
))

}
