% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_opendap.R
\name{getOdapOptParam}
\alias{getOdapOptParam}
\title{Get OPeNDAP optional parameters}
\usage{
getOdapOptParam(collection, roi, loginCredentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{roi}{object of class \code{sf} or \code{sfc}. Region of interest}

\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
a list with the following named objects :
\itemize{
 \item{*roiSpatialIndexBound*: }{OPeNDAP indices for the spatial coordinates of the bounding box of the ROI (minLat, maxLat, minLon, maxLon)}
 \item{*availableVariables*: }{Variables available for the collection of interest}
 \item{*roiSpatialBound*: }{The spatial coordinates of the bounding box of the ROI expressed in the CRS of the collection}
 \item{*OpenDAPtimeVector*: }{The time vector, or NULL if the collection does not have a time vector}
}
}
\description{
Get the OPeNDAP optional parameters to further provide as input \code{optionalsOpendap} parameter of the \code{getUrl} function
}
\details{
When it is needed to loop the function \code{getUrl} over several time frames, it is advised to previously run the function \code{getOdapOptParam} and provide the output as input \code{optionalsOpendap} parameter of the \code{getUrl} function.
This will save much time, as internal parameters will be calculated only once.
}
\examples{

\dontrun{
require(sf)
roi <- sf::st_read(system.file("extdata/roi_example.gpkg", package = "opendapr"),quiet=TRUE)
earthdata_username="user"
earthdata_password="pass"
login_earthdata(c(earthdata_username,earthdata_password))

optOdap_mod11a1 <- getOdapOptParam("MOD11A1.006",roi)
}
}
