% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptParam.R
\name{getOptParam}
\alias{getOptParam}
\title{Precompute the parameter \code{optParam} of the function \link{getUrl}}
\usage{
getOptParam(collection, roi, loginCredentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{roi}{object of class \code{sf} or \code{sfc} Region of interest. Must be POLYGON-type geometry. Can be composed of several features (see details)}

\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
a list with the following named objects :
\itemize{
 \item{*roiSpatialIndexBound*: }{OPeNDAP indices for the spatial coordinates of the bounding box of the ROI (minLat, maxLat, minLon, maxLon)}
 \item{*availableVariables*: }{Variables available for the collection of interest}
 \item{*roiSpatialBound*: }{The spatial coordinates of the bounding box of the ROI expressed in the CRS of the collection}
 \item{*OpenDAPXVector*: }{The X (longitude) vector}
 \item{*OpenDAPYVector*: }{The Y (longitude) vector}
 \item{*OpenDAPtimeVector*: }{The time vector, or NULL if the collection does not have a time vector}
 \item{*modis_tile*: }{The MODIS tile(s) number(s) for the ROI or NULL if the collection is not MODIS}
}
}
\description{
Precompute the parameter \code{optParam} to further provide as input of the \link{getUrl} function. Useful to speed-up the overall processing time.
}
\details{
When it is needed to loop the function \link{getUrl} over several time frames, it is advised to previously run the function \code{getOptParam} and provide the output as input \code{optParam} parameter of the \link{getUrl} function.
This will save much time, as internal parameters will be calculated only once.
}
\examples{

require(sf)

# Login to Earthdata
earthdata_credentials<-readLines("/home/ptaconet/opendapr/.earthdata_credentials.txt")
earthdata_username=strsplit(earthdata_credentials,"=")[[1]][2]
earthdata_password=strsplit(earthdata_credentials,"=")[[2]][2]
login<-login_earthdata(c(earthdata_username,earthdata_password))

# Get the optional parameters for the collection MOD11A1.006 and the roi :
roi <- sf::st_read(system.file("extdata/roi_example.gpkg", package = "opendapr"),quiet=TRUE)
(optParam_mod11a1 <- getOptParam("MOD11A1.006",roi) )

}
