% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{getSRTMtileNames}
\alias{getSRTMtileNames}
\title{Get SRTM tile(s) intersecting a given ROI}
\usage{
getSRTMtileNames(roi)
}
\arguments{
\item{roi}{object of class \code{sf} or \code{sfc}. Region of interest}
}
\value{
a character string vector with the SRTM tiles intersecting the ROI
}
\description{
Get SRTM tile(s) intersecting a given ROI
}
