% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl.R
\name{getUrl}
\alias{getUrl}
\title{Get the URL of}
\usage{
getUrl(collection, variables, roi, timeRange, outputFormat = "nc4",
  singleNetcdf = TRUE, optionalsOpendap = NULL,
  loginCredentials = NULL, verbose = FALSE)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{variables}{string vector. Variables to retrieve for the collection of interest}

\item{roi}{sf POINT or POLYGON. Region of interest}

\item{timeRange}{date(s) / POSIXlt of interest (single date/datetime or time frame) (see details)}

\item{outputFormat}{string. Output format. Available choices : "nc4" (default), "ascii", "",}

\item{singleNetcdf}{boolean. Get the URL either as a single netcdf file that encompasses the whole time frame (TRUE) or as multiple files (1 for each date) (FALSE). Default to TRUE}

\item{optionalsOpendap}{list of optional arguments (see details)}

\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}

\item{verbose}{boolean. Verbose (default FALSE)}
}
\value{
a data.frame with one row for each dataset and 3 columns  :
 \itemize{
 \item{*time_start*: }{Start Date/time for the dataset}
 \item{*name*: }{An indicative name for the dataset}
 \item{*url*: }{URL of the dataset}
 \item{*destfile*: }{An indicative destination file for the dataset}
 }
}
\description{

}
\details{

}
\note{

}
\examples{





}
