% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to download data}
\usage{
login(credentials, source, verbose = TRUE)
}
\arguments{
\item{credentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}

\item{source}{source. Currently available: Earthdata for EOSDIS Earthdata Login with usual credentials}

\item{verbose}{boolean. Verbose (default FALSE)}
}
\value{
None.
}
\description{
In case a login is required, login before querying data servers
}
\details{
Earthdata login enables to download MODIS, VNP, SMAP, GPM data.

Create an account to Earthdata here : \url{https://urs.earthdata.nasa.gov/}.
}
\examples{

\donttest{
username <- Sys.getenv("earthdata_un")
password <- Sys.getenv("earthdata_pw")
login(c(username,password),"earthdata")
}

}
