% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_earthdata}
\alias{login_earthdata}
\title{Login to Earthdata}
\usage{
login_earthdata(loginCredentials)
}
\arguments{
\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
Invisible object with login to earthdata. Login will no longer be needed for the current session.
}
\description{
Login to Earthdata with usual credentials
}
\details{
To open an Earthdata account go to : https://urs.earthdata.nasa.gov/profile
}
\examples{

\dontrun{
earthdata_username="user"
earthdata_password="pass"
login_earthdata(c(earthdata_username,earthdata_password))
}

}
