% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_usgs}
\alias{login_usgs}
\title{Login to USGS EROS ERS}
\usage{
login_usgs(login_credentials, verbose = TRUE)
}
\arguments{
\item{login_credentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}

\item{verbose}{boolean. Verbose (default FALSE)}
}
\value{
None.
}
\description{
Login to USGS EROS Registration System (ERS) with usual credentials
}
\details{
Open an account here : \url{https://ers.cr.usgs.gov/register/}.
}
\note{
The function is inspired from : \url{https://github.com/16EAGLE/getSpatialData/blob/master/R/gSD_login.R}
}
\examples{

\donttest{
username <- Sys.getenv("usgs_un")
password <- Sys.getenv("usgs_pw")
login_usgs(c(username,password))
}

}
