% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisfast-package.R
\docType{package}
\name{modisfast-package}
\alias{modisfast}
\alias{modisfast-package}
\title{modisfast: Fast and Efficient Access to MODIS Earth Observation Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programmatic interface to several NASA Earth Observation 'OPeNDAP' servers (Open-source Project for a Network Data Access Protocol) (\url{https://www.opendap.org/}). Allows for easy downloads of MODIS subsets, as well as other Earth Observation datacubes, in a time-saving and efficient way : by sampling it at the very downloading phase (spatially, temporally and dimensionally).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ptaconet/modisfast}
  \item Report bugs at \url{https://github.com/ptaconet/modisfast/issues}
}

}
\author{
\strong{Maintainer}: Paul Taconet \email{paul.taconet@gmail.com} (\href{https://orcid.org/0000-0001-7429-7204}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul Taconet \email{paul.taconet@gmail.com} (\href{https://orcid.org/0000-0001-7429-7204}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Nicolas Moiroux \email{nicolas.moiroux@ird.fr} (\href{https://orcid.org/0000-0001-6755-6167}{ORCID}) [funder]
  \item French National Research Institute for Sustainable Development, IRD [funder]
}

}
\keyword{internal}
