% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odr_login.R
\name{odr_login}
\alias{odr_login}
\title{Login to query servers and download data}
\usage{
odr_login(credentials, source, verbose = TRUE)
}
\arguments{
\item{credentials}{vector string of length 2 with username and password. optional.}

\item{source}{source. See details}

\item{verbose}{boolean. optional. Verbose (default TRUE)}
}
\value{
None.
}
\description{
Login before querying data servers
}
\details{
Current options for parameter \code{"source"} are :
\itemize{
\item{"earthdata"}{ : to query and download MODIS, VNP, SMAP, GPM collections}
}
Create an account to Earthdata here : \url{https://urs.earthdata.nasa.gov/}.
}
\examples{
# odr_login to Earthdata
\donttest{
username <- Sys.getenv("earthdata_un")
password <- Sys.getenv("earthdata_pw")
odr_login(credentials = c(username,password),source = "earthdata")
}

}
