% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendapr-package.R
\docType{package}
\name{opendapr-package}
\alias{opendapr}
\alias{opendapr-package}
\title{opendapr: Harmonized and Optimized Download of Various Kinds of Spatio-Temporal Earth Science Data in R using the OPeNDAP Framework}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides functions to download various kinds of spatial-temporal Earth science datacubes in a time-saving and efficient way : by sampling it at the very downloading phase (spatially, temporally and dimensionally) using the OPeNDAP framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ptaconet/opendapr}
  \item Report bugs at \url{https://github.com/ptaconet/opendapr/issues}
}

}
\author{
\strong{Maintainer}: Paul Taconet \email{paul.taconet@ird.fr} (0000-0001-7429-7204)

}
\keyword{internal}
