% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download several datasets given their URLs and a destination files}
\usage{
download_data(df_to_dl, parallel = FALSE, credentials = NULL,
  source = NULL, verbose = FALSE)
}
\arguments{
\item{df_to_dl}{data.frame. Urls and destination files of dataset to download. Typically output of \link{get_url}. See Details for the structure}

\item{parallel}{boolean. Parallelize the download ? Default to FALSE}

\item{credentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}

\item{source}{source. Currently available: Earthdata for EOSDIS Earthdata Login with usual credentials}

\item{verbose}{boolean. Verbose (default FALSE)}
}
\value{
a data.frame with the same structure of the input data.frame \code{df_to_dl} + columns providing details of the data downloaded. The additional olumns are :
\itemize{
\item{"fileDl": }{Booloean (dataset downloaded or failure)}
\item{"dlStatus": }{Download status : 1 = download ok ; 2 = download error ; 3 = dataset was already existing in destination file }
\item{"fileSize": }{file size on disk}
}
}
\description{
This function enables to download datasets, enventually parallelizing the download. In a data import workflow, this function is typically used after a call to the \link{get_url} function. The output value of \link{get_url} can be used as input of parameter \code{df_to_dl} of the \link{download_data} function.
}
\details{
Parameter \code{df_to_dl} is a data.frame with the following minimal structure :
\itemize{
\item{"url": }{URL of the source file (character string)}
\item{"destfile": }{Destination file (character string)}
}

Parameter \code{source} takes NULL as default value. Options are :
\itemize{
\item{ \code{NULL} : } {when no login is required to download the data). }
\item{ \code{"earthdata"} : } {to download data requiring a login to Earthdata }
}
}
