% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optional_parameters.R
\name{get_optional_parameters}
\alias{get_optional_parameters}
\title{Precompute the parameter \code{opt_param} of the function \link{get_url}}
\usage{
get_optional_parameters(collection, roi, credentials = NULL)
}
\arguments{
\item{collection}{string. mandatory. Collection of interest (see details of \link{get_url}).}

\item{roi}{object of class \code{sf} or \code{sfc}. mandatory. Region of interest. Must be POLYGON-type geometry. Can be composed of several features (see details of \link{get_url}).}

\item{credentials}{vector string of length 2 with username and password. optional.}
}
\value{
a list with the following named objects :
\describe{
 \item{roiSpatialIndexBound}{OPeNDAP indices for the spatial coordinates of the bounding box of the ROI (minLat, maxLat, minLon, maxLon)}
 \item{availableVariables}{Variables available for the collection of interest}
 \item{roiSpatialBound}{}
 \item{roiSpatialBound }{The spatial coordinates of the bounding box of the ROI expressed in the CRS of the collection}
 \item{OpenDAPXVector}{The X (longitude) vector}
 \item{OpenDAPYVector}{The Y (longitude) vector}
 \item{OpenDAPtimeVector}{The time vector, or NULL if the collection does not have a time vector}
 \item{modis_tile}{The MODIS tile(s) number(s) for the ROI or NULL if the collection is not MODIS}
}
}
\description{
Precompute the parameter \code{opt_param} to further provide as input of the \link{get_url} function. Useful to speed-up the overall processing time.
}
\details{
When it is needed to loop the function \link{get_url} over several time frames, it is advised to previously run the function \code{get_optional_parameters} and provide the output as input \code{opt_param} parameter of the \link{get_url} function.
This will save much time, as internal parameters will be calculated only once.
}
\examples{

\donttest{
require(sf)
require(purrr)

# Login to Earthdata
log <- login(c(Sys.getenv("earthdata_un"),Sys.getenv("earthdata_pw")),source="earthdata")

# Get the optional parameters for the collection MOD11A1.006 and the roi :
roi <- st_as_sf(data.frame(
geom="POLYGON ((-5.82 9.54, -5.42 9.55, -5.41 8.84, -5.81 8.84, -5.82 9.54))"),
wkt="geom",crs = 4326)

opt_param_mod11a1 <- get_optional_parameters("MOD11A1.006",roi)
str(opt_param_mod11a1)

# Now we can provide opt_param_mod11a1 as input parameter of the function get_url().

time_ranges <- list(as.Date(c("2016-01-01","2016-01-31")),
                   as.Date(c("2017-01-01","2017-01-31")),
                   as.Date(c("2018-01-01","2018-01-31")),
                   as.Date(c("2019-01-01","2019-01-31")))

(urls_mod11a1 <- map(.x = time_ranges, ~get_url(
 collection = "MOD11A1.006",
 variables = c("LST_Day_1km","LST_Night_1km","QC_Day","QC_Night"),
 roi = roi,
 time_range = .x,
 opt_param = opt_param_mod11a1)
))

}
}
