% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_info.R
\name{get_variables_info}
\alias{get_variables_info}
\title{Get informations related to the variables available for a given collection}
\usage{
get_variables_info(collection, login_credentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{login_credentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
A data.frame with the available variables for the collection, and a set of related information for each variable.
The variables marked as "extractable" in the column *extractable_w_opendapr* can be provided as input parameter \code{variables} of the function \link{get_url}
}
\description{
Outputs a data frame with the variables available for a given collection
}
\examples{

\dontrun{
# login to usgs
login_usgs(c("my.usgs.username","my.usgs.password"))

# Get the collections implemented in the package :
collections_available <- get_collections_available()

# Get the variables available for the collection MOD11A1.006
(df_varinfo<-get_variables_info("MOD11A1.006"))
}

}
