% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollectionInfo.R
\name{getCollectionInfo}
\alias{getCollectionInfo}
\title{Get information on a collection of interest}
\usage{
getCollectionInfo(collection)
}
\value{
Opens a browser window with the metadata of the collection
}
\description{
Opens a browser window with the metadata of the collection
}
\examples{

# Get the collections implemented :
opendapr:::opendapMetadata_internal$collection

getCollectionInfo("MOD11A1.006")

}
