% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariablesInfo.R
\name{getVariablesInfo}
\alias{getVariablesInfo}
\title{Get variables information for a collection of interest}
\usage{
getVariablesInfo(collection, loginCredentials = NULL)
}
\arguments{
\item{collection}{string. Collection of interest.}

\item{loginCredentials}{vector string. In case of data that needs login : string vector of length 2 with username and password}
}
\value{
A data.frame with the available variables for the collection, and a set of related information
}
\description{
Opens a browser window with the metadata of the collection
}
\examples{

\dontrun{

# login
earthdata_username="user"
earthdata_password="pass"
login<-login_earthdata(c(earthdata_username,earthdata_password))

# Get the collections implemented :
opendapr:::opendapMetadata_internal$collection

df_varinfo<-getVariablesInfo("MOD11A1.006"))
View(df_varinfo)

}

}
