% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_earthdata}
\alias{login_earthdata}
\title{Login to earthdata}
\usage{
login_earthdata(loginCredentials)
}
\arguments{
\item{loginCredentials}{vector string of length 2 with username and password}
}
\value{
Invisible object with login to earthdata. Login will no longer be needed for the current sessions.
}
\description{
Login to earthdata
}
\examples{

\dontrun{
earthdata_username="user"
earthdata_password="pass"
login<-login_earthdata(c(earthdata_username,earthdata_password))
}

}
