% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_import_data.R
\name{mf_import_data}
\alias{mf_import_data}
\title{Import the time series in R as a \code{SpatRaster} object}
\usage{
mf_import_data(
  path,
  collection_source,
  output_class = "SpatRaster",
  proj_epsg = NULL
)
}
\arguments{
\item{path}{string character vector. mandatory. The path to the local directory where the data are stroed.}

\item{collection_source}{character string. mandatory. The collection source (one of "MODIS", "VIIRS", "GPM")}

\item{output_class}{character string. Output object class. Currently only "SpatRaster" implemented.}

\item{proj_epsg}{numeric. EPSG of the desired projection for the output raster (default : source projection for the data)}
}
\description{
Import a time series as a \code{SpatRaster}object
}
\note{
Reprojecting using the argument \code{proj_epsg} might take long over large ROIs.
}
\examples{

\donttest{

require(sf)
require(magrittr)
require(terra)

username <- "earthdata_un"
password <- "earthdata_pw"
log <- mf_login(credentials = c(username,password))

roi <- st_as_sf(data.frame(
id = "roi_test",
geom="POLYGON ((-5.82 9.54, -5.42 9.55, -5.41 8.84, -5.81 8.84, -5.82 9.54))"),
wkt="geom",crs = 4326)


time_range = as.Date(c("2017-01-01","2017-01-30"))

urls_mod11a1 <- mf_get_url(collection = "MOD11A1.061",
variables = c("LST_Day_1km","LST_Night_1km"),
roi = roi,
roi_id = roi_id,
time_range = time_range
)

res_dl <- mf_download_data(urls_mod11a1)

## import as terra::SpatRast

modis_ts <- mf_import_data(dirname(res_dl$destfile[1]), collection_source = "MODIS")

plot(modis_ts)

}
}
