% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_login.R
\name{mf_login}
\alias{mf_login}
\title{Login to query servers and download data}
\usage{
mf_login(credentials, source = "earthdata", verbose = TRUE)
}
\arguments{
\item{credentials}{vector string of length 2 with username and password. optional.}

\item{source}{source. See details}

\item{verbose}{boolean. optional. Verbose (default TRUE)}
}
\value{
None.
}
\description{
Login before querying data servers
}
\details{
Current options for parameter \code{"source"} are :
\itemize{
\item{"earthdata"}{ : to query and download MODIS, VNP, SMAP, GPM collections}
}
Create an account to Earthdata here : \url{https://urs.earthdata.nasa.gov/}.
}
\examples{

\donttest{
username <- "earthdata_un"
password <- "earthdata_pw"
mf_login(credentials = c(username,password))
}

}
