% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisfast-package.R
\docType{package}
\name{modisfast-package}
\alias{modisfast}
\alias{modisfast-package}
\title{modisfast: Fast and efficient access to MODIS data in R (and several other Earth datacubes sources)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to download MODIS and several other Earth datacubes sources in a time-saving and efficient way : by sampling it at the very downloading phase (spatially, temporally and dimensionally) through the OPeNDAP framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ptaconet/modisfast}
  \item Report bugs at \url{https://github.com/ptaconet/modisfast/issues}
}

}
\author{
\strong{Maintainer}: Paul Taconet \email{paul.taconet@ird.fr} (\href{https://orcid.org/0000-0001-7429-7204}{ORCID})

}
\keyword{internal}
